requirements.txt
streamlit>=1.34.0
pandas>=2.2.0
openpyxl>=3.1.2
PyYAML>=6.0.1
pydantic>=2.7.0
xlsxwriter>=3.2.0
pytest>=8.2.0

